requirements.txt
Pillow===9.4.0
python-decouple===3.8
aiogram===3.1.1
